Support for transient version vertex state.
When a versioned vertex is created in transaction scope, it has no
verson yet, this is because the next version is determined only when the transaction
is committed.

This causes errors when the vertex properties or associated edges are retrieved
as the version does not exist yet and required during content retrieval.

One solution to solve this is by setting the forVersion vertex property to the latest
graph version, but this concept is not 100% accurate because the version is actually not the latest
but the next version to be committed,

Thus, currently this patch solves this problem by flagging the created vertex as transient,
when the vertex is versioned (persisted) this flag is cleaned. the default value of this flag is false of course.

Any try to retrieve data on a transient vertex will be blocked and an exception will be thrown.

Tests in transactional graph were added for transient vertices as well.