Bug in getVertex(Object id) that retrieves deleted vertices.
Vertices never gets deleted, deleting a vertex only sets a limit of the end
version range of the vertex.

Invoking getVertex(Object id) should not just look for the vertex by ID,
as even if it gets deleted, it will exist, but must pass through
the logic of finding if the vertex is valid for a certain version (getVertexForVersion(Vertex vertex, V version)).